Added Src file.
Created a folder of common/styles/font and added a custom font - Gilroy. Inside styles added variable.css for adding background and text colour to my root file. In the Pages folder added homePage.js which will be rendered to App.js so that no modifications are required in App.js. Created my first component as HeroSection to make the landing page.